// client/obras/src/components/FormConstruction/Interface.ts
export interface IFormData {
  nome?: string;
  anoConstrucao?: number;
  coordenacao?: string;
  gerencia?: string;
  diretoria?: string;
  outorga?: string;
  titularidade?: string;
  cidade?: string;
  estado?: string;
  latitude?: string;
  longitude?: string;
  tipo?: string;
  risco?: string;
}
export interface NewFormData {
  nome?: string;
  anoConstrucao?: number;
  coordenacao?: string;
  gerencia?: string;
  diretoria?: string;
  outorga?: string;
  titularidade?: string;
}